language changes: integer type fallback is gone; + on String is back.
also features an updated version of `lex!` which allows for
user-defined parsers (at expense of a slight syntax modification);
we need it as, say, `f32` and `f64` can't be simply unified to
the same code without introducing a new trait.
a new trait `FromStrPrefix` subsumes the existing `scan_*` extension
methods, and thus `util::std::str` has been greatly reduced.

also due to the recent lifetime inference problem (rust-lang/rust#15140)
we now use custom methods `.as_ref_slice()` and `.as_ref_slice_or()`
for the major source of this problem.

+++ src/gfx/skin/parse.rs
@@ -234,8 +234,6 @@ impl FromJson for Gradient {
 
 impl FromJson for Expr {
     fn from_json(json: Json) -> Result<Expr,String> {
-        use util::std::str::StrUtil;
-
         let expr = match json {
             String(s) => s,
             Number(v) => { return Ok(ENum(v as f32)); }
@@ -253,7 +251,7 @@ impl FromJson for Expr {
         let mut minus = false;
         loop {
             let mut v = 0.0;
-            if !lex!(s; ws*, f64 -> v, str* -> s, !) { invalid!("expected a number"); }
+            if !lex!(s; ws*, f32 -> v, str* -> s, !) { invalid!("expected a number"); }
             if minus { v = -v; }
             if s.starts_with("%") {
                 v /= 100.0;
